gk: handle hs_hash_add_key_with_hash() == -EEXIST
In rare situations, hs_hash_add_key_with_hash() will try to add
a flow already in the flow table. Callers of
hs_hash_add_key_with_hash() ignored these cases and discarded
the associate updates.

The reason for not handling those cases was that
rte_hash_add_key_with_hash() does not return the entry ID when
it already exists in the table, and adding a lookup to
the hot path was not desirable. However,
hs_hash_add_key_with_hash(), from the new hash table library,
provides the entry ID when it is already in the table.