lib/net: monitor state of the links
Having log entries reporting changes in the state of the links
eases diagnoses in production. This commit adds these log entries
through the Link State Change (LSC) interuption that most NICs
support. The information that a given NIC does not support
the LSC interuption is logged, but its link state is not monitored.

Since the callbacks of interuptions run in threads that are not
logical cores, this commit introduces the macro MAIN_LOG() to
Gatekeeper's log library to be used in these contexts.