sol: require parameter req_channel_bw_mbps
The original configuration of SOL blocks assumes that
the bandwidth of the back interface of a Gatekeeper server
is representative of the bandwidth of the destination network.
This assumption does not hold in production deployments of
Gatekeeper.  For example, Gatekeeper servers typically operate
in networks with higher speeds (e.g., 40Gbps) than the protected
destination (e.g., 10Gbps).

This commit introduces the parameter destination_bw_gbps to address
this reality.