gatekeeper: replace gettid() with rte_gettid()
Not only does gettid() require #define _GNU_SOURCE, but it has also
presented header problems in different versions of Lib C. So this
commit eases the effort to compile Gatekeeper on different distros.

A small benefit gained from using rte_gettid() is that it caches
the thread ID in a per-thread variable to avoid subsequent system
calls.